+++ src/user/user.cpp
@@ -12,9 +12,9 @@
  * @param nickname
  * @param type
  */
-User::User(size_t age, std::string name, std::string nickname, enum UserTypes type) : age(age), name(name), nickname(nickname), type(type){}
+User::User(unsigned int age, std::string name, std::string nickname, enum UserTypes type) : age(age), name(name), nickname(nickname), type(type){}
 
-size_t User::getAge() const{
+unsigned int User::getAge() const{
     return age;
 }
